# dist/swagger.json
{
	"components": {
...
		"schemas": {
			"Pick_ICountryType.Exclude_keyofICountryType.flag-or-countryId-or-status__": {
				"properties": {
					"index": {
						"type": "string"
					},
					"countrycode": {
						"type": "string"
					},
					"designation": {
						"type": "string"
					}
				},
				"required": [
					"index",
					"countrycode",
					"designation"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ICountryType.flag-or-countryId-or-status_": {
				"$ref": "#/components/schemas/Pick_ICountryType.Exclude_keyofICountryType.flag-or-countryId-or-status__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Pick_ICountryType.Exclude_keyofICountryType.flag__": {
				"properties": {
					"countryId": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"index": {
						"type": "string"
					},
					"countrycode": {
						"type": "string"
					},
					"designation": {
						"type": "string"
					}
				},
				"required": [
					"index",
					"countrycode",
					"designation"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ICountryType.flag_": {
				"$ref": "#/components/schemas/Pick_ICountryType.Exclude_keyofICountryType.flag__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Pick_ICountryType.countryId_": {
				"properties": {
					"countryId": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Pick_IEtatType.Exclude_keyofIEtatType.etatId-or-status__": {
				"properties": {
					"designation": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"designation",
					"country"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IEtatType.etatId-or-status_": {
				"$ref": "#/components/schemas/Pick_IEtatType.Exclude_keyofIEtatType.etatId-or-status__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Pick_IEtatType.Exclude_keyofIEtatType.status__": {
				"properties": {
					"designation": {
						"type": "string"
					},
					"etatId": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"designation",
					"country"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IEtatType.status_": {
				"$ref": "#/components/schemas/Pick_IEtatType.Exclude_keyofIEtatType.status__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Pick_IEtatType.etatId_": {
				"properties": {
					"etatId": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "uptodate-ekivalApp-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/admin/contries": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"country": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"country",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Contries"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_ICountryType.flag-or-countryId-or-status_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Contries",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"countries": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"countries",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Contries"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"country": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"country",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Contries"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_ICountryType.flag_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Contries"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_ICountryType.countryId_"
							}
						}
					}
				}
			}
		},
		"/api/admin/etats": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"etat": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"etat",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Etats"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IEtatType.etatId-or-status_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Etats",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"etats": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"etats",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Etats"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"etat": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"etat",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Etats"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IEtatType.status_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Etats"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_IEtatType.etatId_"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/v1"
		}
	]
}